% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_scholar.R
\name{get_fields}
\alias{get_fields}
\title{A comma-separated list of the fields available for use in API calls}
\format{
The following list of the fields to be returned are recognized:
\itemize{
\item paper_search : fields on searching for papers by keyword.
\item paper : fields on looking up details about a paper.
\item paper_authors : fields on looking up details about a paper's authors.
\item paper_citations : fields on fetching details about the papers the cite this paper (i.e. papers in whose bibliography this paper appears).
\item paper_references : fields on fetching details about the papers cited by this paper (i.e. appearing in this paper's bibliography).
\item author : fields on looking up details about an author.
\item author_papers : fields on fetching the papers of an author in batches,only retrieves the most recent 10,000 citations/references for papers belonging to the batch.
}
}
\usage{
get_fields()
}
\value{
list representing fields parms.
}
\description{
This function get fields for fetching paper and author data from the Semantic Scholar corpus.
}
\details{
See for website \url{https://api.semanticscholar.org/graph/v1} details.
}
\examples{
\dontrun{
# get all fileds
get_fields()

# check type of fileds
names(get_fields())

# get author of fileds
get_fields()[["author"]]
}
}
\author{
Liuyong Ding
}
