% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_scholar.R
\name{lookup_author}
\alias{lookup_author}
\title{Show author profile plot}
\usage{
lookup_author(author_id, open_url = FALSE, plot = TRUE, size = 3, ...)
}
\arguments{
\item{author_id}{\code{string} A plain-text search query string with author identifier.}

\item{open_url}{\code{logical} Default: FALSE. Is open scholar profile homepage.}

\item{plot}{\code{logical} Default: TRUE. Is show scholar profile plot.}

\item{size}{Size of label,default: size = 3.}

\item{...}{Other arguments set by \code{\link{geom_col}}}
}
\value{
a length-2 list of \code{\link{ggplot}} object and raw data.
}
\description{
This function show scholar profile plot
}
\examples{
\dontrun{
res <- lookup_author(author_id = "48980001")
res$plot
}
}
\author{
Liuyong Ding
}
