% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_scholar.R
\name{lookup_paper}
\alias{lookup_paper}
\title{Lookup details about a paper}
\format{
The following types of paper_id are supported:
\itemize{
\item <sha> - a Semantic Scholar ID, e.g. 649def34f8be52c8b66281af98ae884c09aef38b
\item CorpusId:<id> - Semantic Scholar numerical ID, e.g. 215416146
\item DOI:<doi> - a Digital Object Identifier, e.g. DOI:10.18653/v1/N18-3011
\item ARXIV:<id> - arXiv.rg, e.g. ARXIV:2106.15928
\item MAG:<id> - Microsoft Academic Graph, e.g. MAG:112218234
\item ACL:<id> - Association for Computational Linguistics, e.g. ACL:W12-3903
\item PMID:<id> - PubMed/Medline, e.g. PMID:19872477
\item PMCID:<id> - PubMed Central, e.g. PMCID:2323736
\item URL:<url> - URL from one of the sites listed below, e.g. URL:https://arxiv.org/abs/2106.15928v1
\itemize{
\item semanticscholar.org \url{https://www.semanticscholar.org}
\item arxiv.org \url{https://arxiv.org/}
\item aclweb.org \url{https://www.aclweb.org/}
\item acm.org \url{https://www.acm.org/}
\item biorxiv.org \url{https://www.biorxiv.org/}
}
}

The following returned results are recognized:
\itemize{
\item paperId \code{string}
\item externalIds \code{object} Other catalog IDs for this paper, if known. Supports ArXiv, MAG, ACL, PubMed, Medline, PubMedCentral, DBLP, DOI.
\item url \code{string} URL on the Semantic Scholar website
\item title \code{string} This field will be provided if no fields are specified.
\item abstract \code{string}
\item venue \code{string}
\item year \code{integer}
\item referenceCount \code{integer}
\item citationCount \code{integer}
\item influentialCitationCount \code{integer} \url{https://www.semanticscholar.org/faq#influential-citations}
\item isOpenAccess \code{boolean} \url{https://www.openaccess.nl/en/what-is-open-access}
\item fieldsOfStudy \code{object} A list of high-level academic categories.
\item authors \code{object} 	Array of objects (Author Info).
\item citations \code{object} Array of objects (Paper Info)
\item references \code{object} Array of objects (Paper Info
\item embedding \code{object} 	object (Embedding)
\item tldr \code{object} Too Long; Didn't Read. Artificial Intelligence \url{https://www.semanticscholar.org/product/semantic-reader} summarizes academic articles for you.
}
}
\usage{
lookup_paper(
  paper_id,
  fields = c("externalIds", "title", "abstract", "year", "tldr")
)
}
\arguments{
\item{paper_id}{\code{string} Supported paper identifier services: Semantic Scholar ID, Semantic Scholar numerical ID, Digital Object Identifier, arXiv.rg, Microsoft Academic Graph, Association for Computational Linguistics, PubMed/Medline, PubMed Central and  URL from one of the sites listed below. See Format for detalis.}

\item{fields}{\code{string} A comma-separated list of the fields to be returned, support one or more fileds, for example fileds = get_fields()[['papers']] was all fields to be returned. See Format or see \code{\link{get_fields}} to set more params.}
}
\value{
list representing paper objects
}
\description{
This function fetch a paper data from the Semantic Scholar corpus.
}
\details{
See website \url{https://api.semanticscholar.org/graph/v1#operation/get_graph_get_paper} for details.
}
\examples{
\dontrun{
res <- lookup_paper(paper_id = "649def34f8be52c8b66281af98ae884c09aef38b")
}
}
\author{
Liuyong Ding
}
