% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2miner-package.R
\name{S2miner-package}
\alias{S2miner-package}
\alias{S2miner}
\title{Semantic Scholar Mining with R}
\description{
Semantic Scholar \url{https://www.semanticscholar.org} is a free, AI-powered research tool for scientific literature, aimed to help scholars discover new insights.
This package helps users to extract, analyze and visualize scholarly data from ‘Semantic Scholar’ API, thus providing new insight into bibliometrics study.
}
\references{
\itemize{
\item Waleed Ammar et al. 2018. Construction of the Literature Graph in Semantic Scholar. NAACL \url{https://www.semanticscholar.org/paper/09e3cf5704bcb16e6657f6ceed70e93373a54618}
\item Markus Skyttner (2021). semanticscholar: Data From Semantic Scholar Leightweight API. R package version 0.2.0. \url{https://github.com/KTH-Library/semanticscholar}
\item Najko Jahn (2021). semscholar: Interface to the Semantic Scholar search engine. R package version 0.1.0. \url{https://github.com/njahn82/semscholar}
}
}
\keyword{package}
