% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_scholar.R
\name{search_papers}
\alias{search_papers}
\title{Searching for papers by keyword}
\format{
The following case-sensitive paper fields are recognized:
\itemize{
\item paperId : S2 generated research paper ID (Always included).
\item externalIds : Other catalog IDs for this paper, if known. Supports ArXiv, MAG, ACL, PubMed, Medline, PubMedCentral, DBLP, DOI.
\item title : Research paper title (Included if no fields are specified).
\item url : URL on the Semantic Scholar website.
\item abstract : Extracted abstract of the paper.
\item venue : Extracted publication venue for this paper.
\item year : Year this paper was published as integer.
\item referenceCount : The number of reference for this paper.
\item citationCount : The number of citation for this paper.
\item influentialCitationCount : The number of influential citation for this paper.
\item isOpenAccess : \url{https://www.openaccess.nl/en/what-is-open-access}
\item fieldsOfStudy : A list of high-level academic categories.
\item authors : Author(s) for this paper (Up to 500 will be returned).
\itemize{
\item authorId : Always included
\item name : Always included
}
}
}
\usage{
search_papers(keyword, offset = 0, limit = 100, fields = NULL)
}
\arguments{
\item{keyword}{\code{string} A plain-text search query string. No special query syntax is supported.Supports AND and OR and inclusion and exclusion of terms, for example "title=bibliometrics" or "+Epidemic+Modeling+Canada-COVID".See blog post \url{https://medium.com/ai2-blog/building-a-better-search-engine-for-semantic-scholar-ea23a0b661e7} for a description of our search relevance algorithm.}

\item{offset}{\code{integer} Default: 0. When returning a list of results, start with the element at this position in the list.}

\item{limit}{\code{integer} Default: 100. The maximum number of results to return. The sum of offset and limit must be < 10000.}

\item{fields}{\code{string} A comma-separated list of the fields to be returned, support one or more fileds, for example fileds = get_fields()[['paper_search']] was all fields to be returned. See Format or see \code{\link{get_fields}} to set more params}
}
\value{
list representing paper objects
}
\description{
This function retrieves Semantic Scholar \url{https://www.semanticscholar.org} data for a keyword search.
}
\details{
You can use the Semantic Scholar API endpoint up to 100 requests per 5 minutes to test your application idea. To access a higher rate limit, complete the form \url{https://www.semanticscholar.org/product/api#Partner-Form} to request authentication for your project.
Current API users may have noticed that some paper requests would error out if the paper data was too large. The new service avoids this problem by, first, allowing you to specify exactly the data you’re interested in and, second, allowing you to paginate through large result sets, such as papers with many citations.
}
\examples{
\dontrun{
res <- search_papers(keyword = "bibliometrics",fields = c("title","abstract","authors"))
}
}
\author{
Liuyong Ding
}
