% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_scholar.R
\name{traverse_author}
\alias{traverse_author}
\title{Traversing the literature graph of an author's papers}
\usage{
traverse_author(author_id, offset = 0, limit = 100, fields = NULL)
}
\arguments{
\item{author_id}{\code{string} A plain-text search query string with author identifier.}

\item{offset}{\code{integer} Default: 0. When returning a list of results, start with the element at this position in the list.}

\item{limit}{\code{integer} Default: 100. The maximum number of results to return. The sum of offset and limit must be < 10000.}

\item{fields}{\code{string} A comma-separated list of the fields to be returned, support one or more fileds, for example fileds = get_fields()[['author_papers']] was all fields to be returned. See Format or see \code{\link{get_fields}} to set more params}
}
\value{
list representing author's papers
}
\description{
This function fetch the papers of an author in batches.
}
\details{
Only retrieves the most recent 10,000 citations/references for papers belonging to the batch.
}
\examples{
\dontrun{
res <- traverse_author(author_id = "48980001",fields = c("title","abstract","authors","year"))
}
}
\author{
Liuyong Ding
}
