% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_scholar.R
\name{traverse_paper}
\alias{traverse_paper}
\title{Traversing the literature graph of a paper's information}
\format{
The following types of paper_id are supported:
\itemize{
\item <sha> - a Semantic Scholar ID, e.g. 649def34f8be52c8b66281af98ae884c09aef38b
\item CorpusId:<id> - Semantic Scholar numerical ID, e.g. 215416146
\item DOI:<doi> - a Digital Object Identifier, e.g. DOI:10.18653/v1/N18-3011
\item ARXIV:<id> - arXiv.rg, e.g. ARXIV:2106.15928
\item MAG:<id> - Microsoft Academic Graph, e.g. MAG:112218234
\item ACL:<id> - Association for Computational Linguistics, e.g. ACL:W12-3903
\item PMID:<id> - PubMed/Medline, e.g. PMID:19872477
\item PMCID:<id> - PubMed Central, e.g. PMCID:2323736
\item URL:<url> - URL from one of the sites listed below, e.g. URL:https://arxiv.org/abs/2106.15928v1
\itemize{
\item semanticscholar.org \url{https://www.semanticscholar.org}
\item arxiv.org \url{https://arxiv.org/}
\item aclweb.org \url{https://www.aclweb.org/}
\item acm.org \url{https://www.acm.org/}
\item biorxiv.org \url{https://www.biorxiv.org/}
}
}
}
\usage{
traverse_paper(
  paper_id,
  offset = 0,
  limit = 100,
  type = "citations",
  fields = NULL
)
}
\arguments{
\item{paper_id}{\code{string} Supported paper identifier services: Semantic Scholar ID, Semantic Scholar numerical ID, Digital Object Identifier, arXiv.rg, Microsoft Academic Graph, Association for Computational Linguistics, PubMed/Medline, PubMed Central and  URL from one of the sites listed below. See Format for detalis.}

\item{offset}{\code{integer} Default: 0. When returning a list of results, start with the element at this position in the list.}

\item{limit}{\code{integer} Default: 100. The maximum number of results to return. The sum of offset and limit must be < 10000.}

\item{type}{\code{string} Supported paper query services: "authors","citations", OR "references".}

\item{fields}{\code{string} A comma-separated list of the fields to be returned, support one or more fileds, for example fileds = get_fields()[['paper_citation']] for "citations" was all fields to be returned. See Format or see \code{\link{get_fields}} to set more params.}
}
\value{
list representing paper objects
}
\description{
This function fetch details about the papers the cite this paper (i.e. papers in whose bibliography this paper appears) and cited by this paper (i.e. appearing in this paper's bibliography)
}
\details{
See website \url{https://api.semanticscholar.org/graph/v1#operation/get_graph_get_paper_authors} for details.
}
\examples{
\dontrun{
res <- traverse_paper(paper_id = "649def34f8be52c8b66281af98ae884c09aef38b")
}
}
\author{
Liuyong Ding
}
